fix(dicePool): extract die-face icons to a standalone utility
The activation dialog imported getDieFaceIcon from DicePoolTracker's
co-located state file, which transitively pulls in the actor document
chain (NimbleCharacter -> base.svelte.ts -> ...). That import created
a 10-edge cycle that pnpm circular-deps flagged. Move the icon map
into a dependency-free util in #utils/dicePool/dieFaceIcons.ts and
point both the tracker and the dialog at it.
